# api/microcks-openapi-v1.0.yaml
...
openapi: 3.0.2
info:
  title: Microcks API v1.0
  version: 1.0.0
  description: API offered by Microcks, the mock and testing platform for API and
    microservices (microcks.github.io)
  contact:
    name: Laurent Broudoux
    url: https://github.com/lbroudoux
    email: laurent.broudoux@gmail.com
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
- url: http://microcks.example.com/api
  description: ""
paths:
  /services:
    summary: This path operations deal with Services
    get:
      tags:
      - mock
      parameters:
      - name: page
        description: Page of Services to retrieve (starts at and defaults to 0)
        schema:
          type: integer
        in: query
      - name: size
        description: Size of a page. Maximum number of Services to include in a response
          (defaults to 20)
        schema:
          type: integer
        in: query
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Service'
          description: List of found Services
      security:
      - jwt-bearer:
        - user
      operationId: GetServices
      summary: Get Services and APIs
...
  /services/count:
    get:
      tags:
      - mock
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Counter'
          description: Number of Services in datastore
      security:
      - jwt-bearer:
        - user
      operationId: GetServicesCounter
      summary: Get the Services counter
...
  /services/{id}:
    get:
      tags:
      - mock
      parameters:
      - name: messages
        description: Whether to include details on services messages into result.
          Default is false
        schema:
          type: boolean
        in: query
      responses:
        "200":
          $ref: '#/components/responses/ServiceResponse'
      security:
      - jwt-bearer:
        - user
      operationId: GetService
      summary: Get Service
    delete:
      tags:
      - mock
      responses:
        "200":
          description: Service has been deleted
      security:
      - jwt-bearer:
        - admin
        - manager
      operationId: DeleteService
      summary: Delete Service
      description: Delete a Service
    parameters:
    - name: id
      description: Unique identifier of Service to managed
      schema:
        type: string
      in: path
      required: true
  /artifact/upload:
    summary: Deals with artifacts (SOAP UI project, Postman collection or OpenAPI
      Specification) to be imported by Microcks.
    post:
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/ArtifactUpload'
            examples:
              Artifact upload:
                value: |
                  POST /api/artifact/upload HTTP/1.1
                  Host: microcks
                  User-Agent: curl/7.54.0
                  Accept: */*
                  Authorization: Bearer <REDACTED>
                  Content-Length: 2743
                  Expect: 100-continue
                  Content-Type: multipart/form-data; boundary=------------------------8af8cbb56dd4bde0

                  --------------------------8af8cbb56dd4bde0
                  Content-Disposition: form-data; name="file"; filename="github.json"
                  Content-Type: application/octet-stream

                  THE ARTIFACT HERE

                  --------------------------8af8cbb56dd4bde0--
        required: true
      tags:
      - job
      responses:
        "201":
          content:
            text/plain:
              schema:
                type: string
          description: Artifact was imported and Service found
        "204":
          description: No file attribute found in uploaded data
        "400":
          content:
            text/plain:
              schema:
                type: string
          description: Artifact content is invalid and not understood
      operationId: uploadArtifact
      summary: Upload an artifact
      description: Uploads an artifact (SOAP UI project, Postman collection or OpenAPI
        Specification) to be imported by Microcks.
...
  /tests/service/{serviceId}:
    get:
      tags:
      - test
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TestResult'
          description: List of TestResults for the Service having the requested id
      security:
      - jwt-bearer:
        - user
      operationId: GetTestResultsByService
      summary: Get TestResults by Service
    parameters:
    - name: serviceId
      description: Unique identifier of Service to manage TestResults for
      schema:
        type: string
      in: path
      required: true
...
  /services/{id}/operation:
    put:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OperationOverrideDTO'
        required: true
      tags:
      - mock
      parameters:
      - name: operationName
        description: Name of operation to update
        schema:
          type: string
        in: query
        required: true
      responses:
        "200":
          description: Operation has been updated
        "500":
          description: Operation cannot be updated
      security:
      - jwt-bearer:
        - admin
        - manager
      operationId: OverrideServiceOperation
      summary: Override Service Operation
    parameters:
    - name: id
      description: Unique identifier of Service to managed
      schema:
        type: string
      in: path
      required: true
  /services/{id}/metadata:
    put:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Metadata'
        required: true
      tags:
      - mock
      responses:
        "200":
          description: Service metadata has been updated
        "500":
          description: Update of metadata failed
      security:
      - jwt-bearer:
        - admin
        - manager
      operationId: UpdateServiceMetadata
      summary: Update Service Metadata
    parameters:
    - name: id
      description: Unique identifier of Service to managed
      schema:
        type: string
      in: path
      required: true
  /services/labels:
    get:
      tags:
      - mock
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LabelsMap'
          description: 'Already used labels: keys are label Keys, values are array
            of label Values'
      security:
      - jwt-bearer:
        - admin
      operationId: GetServicesLabels
      summary: Get the already used labels for Services
  /services/search:
    get:
      tags:
      - mock
      parameters:
      - name: queryMap
        description: Map of criterion. Key can be simply 'name' with value as the
          searched string. You can also search by label using keys like 'labels.x'
          where 'x' is the label and value the label value
        schema:
          type: object
          additionalProperties:
            type: string
        in: query
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Service'
          description: List of found Services (filtered according search criteria)
      security:
      - jwt-bearer:
        - user
      operationId: SearchServices
      summary: Search for Services and APIs
components:
  schemas:
    TestCaseResult:
      description: Companion objects for TestResult. Each TestCaseResult correspond
        to a particuliar service operation / action reference by the operationName
        field. TestCaseResults owns a collection of TestStepResults (one for every
        request associated to service operation / action).
      required:
      - success
      - elapsedTime
      - operationName
      properties:
        success:
          description: Flag telling if test case is a success
          type: boolean
        elapsedTime:
          description: Elapsed time in milliseconds since the test case beginning
          type: number
        operationName:
          description: Name of operation this test case is bound to
          type: string
        testStepResults:
          description: Test steps associated to this test case
          type: array
          items:
            $ref: '#/components/schemas/TestStepResult'
    TestRunnerType:
      description: Type of test strategy (different strategies are implemented by
        different runners)
      type: array
      items:
        enum:
        - HTTP
        - SOAP_HTTP
        - SOAP_UI
        - POSTMAN
        - OPEN_API_SCHEMA
        type: string
    Service:
      description: Represents a Service or API definition as registred into Microcks
        repository
      required:
      - name
      - version
      - type
      properties:
        id:
          description: Unique identifier for this Service or API
          type: string
        name:
          description: Distinct name for this Service or API (maybe shared among many
            versions)
          type: string
        version:
          description: Distinct version for a named Service or API
          type: string
        type:
          description: Service or API Type
          enum:
          - REST
          - SOAP_HTTP
          - GENERIC_REST
          - EVENT
          type: string
        operations:
          description: Set of Operations for Service or API
          type: array
          items:
            $ref: '#/components/schemas/Operation'
        xmlNS:
          description: Associated Xml Namespace in case of Xml based Service
          type: string
        metadata:
          $ref: '#/components/schemas/Metadata'
          description: Metadata of Service
...
    ArtifactUpload:
      title: Root Type for ArtifactUpload
      description: |-
        Artifact (SOAP UI project, Postman collection or OpenAPI Specification) to be imported by Microcks.
        This structure represents a mime-multipart file upload (as specified here: https://swagger.io/docs/specification/describing-request-body/file-upload/)
      required:
      - file
      type: object
      properties:
        file:
          format: binary
          description: The artifact to upload
          type: string
    Counter:
      title: Root Type for Counter
      description: The root of the Counter type's schema.
      type: object
      properties:
        counter:
          format: int32
          description: Number of items in a resource collection
          type: integer
      example: |-
        {
            "counter": 12
        }
...
    TestStepResult:
      description: TestStepResult is an entity embedded within TestCaseResult. They
        are created for each request associated with an operation / action of a microservice.
      required:
      - success
      - elapsedTome
      - requestName
      properties:
        success:
          description: Flag telling if test case is a success
          type: boolean
        elapsedTime:
          description: Elapsed time in milliseconds since the test step beginning
          type: number
        requestName:
          description: Name of request this test step is bound to
          type: string
        message:
          description: Error message that may be associated to this test step
          type: string
...
    RequestResponsePair:
      description: Request associated with corresponding Response
      type: object
      allOf:
      - required:
        - request
        - response
        type: object
        properties:
          request:
            $ref: '#/components/schemas/Request'
            description: The request part of the pair
          response:
            $ref: '#/components/schemas/Response'
            description: The Response part of the pair
      - $ref: '#/components/schemas/AbstractExchange'
    TestResult:
      description: Represents the result of a Service or API test run by Microcks.
        Tests are related to a service and made of multiple test cases corresponding
        to each operations / actions composing service. Tests are run against a specific
        endpoint named testedEndpoint. It holds global markers telling if test still
        ran, is a success, how many times is has taken and so on ...
      required:
      - id
      - version
      - testNumber
      - testedEndpoint
      - serviceId
      - success
      - inProgress
      - runnerType
      properties:
        id:
          description: Unique identifier of TestResult
          type: string
        version:
          description: Revision number of this test
          type: number
        testNumber:
          description: Incremental number for tracking number of tests of a service
          type: number
        testedEndpoint:
          description: Endpoint used during test
          type: string
        serviceId:
          description: Unique identifier of service tested
          type: string
        elapsedTime:
          description: Elapsed time in milliseconds since test beginning
          type: number
        success:
          description: Flag telling if test is a success
          type: boolean
        inProgress:
          description: Flag telling is test is still in progress
          type: boolean
        runnerType:
          $ref: '#/components/schemas/TestRunnerType'
          description: Runner used for this test
        testCaseResults:
          description: TestCase results associated to this test
          type: array
          items:
            $ref: '#/components/schemas/TestCaseResult'
    OperationOverrideDTO:
      description: Data Transfer object for grouping the mutable properties of an
        Operation
      type: object
      properties:
        dispatcher:
          description: Type of dispatcher to apply for this operation
          type: string
        dispatcherRules:
          description: Rules of dispatcher for this operation
          type: string
        defaultDelay:
          description: Default delay in milliseconds to apply to mock responses on
            this operation
          type: integer
        parameterConstraints:
          description: Constraints that may apply to incoming parameters on this operation
          type: array
          items:
            $ref: '#/components/schemas/ParameterConstraint'
    ParameterConstraint:
      description: Companion object for Operation that may be used to express constraints
        on request parameters
      required:
      - name
      type: object
      properties:
        name:
          description: Parameter name
          type: string
        required:
          description: Whether it's a required constraint
          type: boolean
        recopy:
          description: Whether it's a recopy constraint
          type: boolean
        mustMatchRegexp:
          description: Whether it's a regular expression matching constraint
          type: string
        in:
          description: Parameter location
          enum:
          - path
          - query
          - header
          type: string
    Metadata:
      description: Commodity object for holding metadata on any entity. This object
        is inspired by Kubernetes metadata.
      required:
      - createdOn
      - lastUpdate
      type: object
      properties:
        createdOn:
          description: Creation date of attached object
          type: integer
        lastUpdate:
          description: Last update of attached object
          type: integer
        annotations:
          description: Annotations of attached object
          type: object
          additionalProperties:
            type: string
        labels:
          description: Labels put on attached object
          type: object
          additionalProperties:
            type: string
    Binding:
      description: Protocol binding details for asynchronous operations
      required:
      - type
      - destinationName
      type: object
      properties:
        type:
          description: Protocol binding identifier
          enum:
          - KAFKA
          type: string
        keyType:
          description: Type of key for Kafka messages
          type: string
        destinationType:
          description: Type of destination for asynchronous messages of this operation
          type: string
        destinationName:
          description: Name of destination for asynchronous messages of this operation
          type: string
    Operation:
      description: An Operation of a Service or API
      required:
      - name
      - method
      type: object
      properties:
        name:
          description: Unique name of this Operation within Service scope
          type: string
        method:
          description: Represents transport method
          type: string
        inputName:
          description: Name of input parameters in case of Xml based Service
          type: string
        outputName:
          description: Name of output parameters in case of Xml based Service
          type: string
        dispatcher:
          description: Dispatcher strategy used for mocks
          type: string
        dispatcherRules:
          description: DispactherRules used for mocks
          type: string
        defaultDelay:
          description: Default response time delay for mocks
          type: number
        resourcePaths:
          description: Paths the mocks endpoints are mapped on
          type: array
          items:
            type: string
        parameterContraints:
          description: Contraints that may apply to mock invocatino on this operation
          type: array
          items:
            $ref: '#/components/schemas/ParameterConstraint'
        bindings:
          description: Map of protocol binding details for this operation
          type: object
          additionalProperties:
            $ref: '#/components/schemas/Binding'
    LabelsMap:
      description: A map which keys are already used labels keys and values are already
        used values for this key
      type: object
      additionalProperties:
        $ref: '#/components/schemas/StringArray'
        type: object
...
    StringArray:
      description: A simple array of String
      type: array
      items:
        type: string
    MessageArray:
      description: Array of Message for Service operations
      type: array
      items:
        $ref: '#/components/schemas/Exchange'
    Exchange:
      oneOf:
      - $ref: '#/components/schemas/RequestResponsePair'
      - $ref: '#/components/schemas/UnidirectionalEvent'
      discriminator:
        propertyName: type
        mapping:
          reqRespPair: '#/components/schemas/RequestResponsePair'
          unidirEvent: '#/components/schemas/UnidirectionalEvent'
...
    ServiceView:
      description: Aggregate bean for grouping a Service an its messages pairs
      required:
      - service
      - messagesMap
      type: object
      properties:
        service:
          $ref: '#/components/schemas/Service'
          description: Wrapped service description
        messagesMap:
          description: Map of messages for this Service. Keys are operation name,
            values are array of messages for this operation
          type: object
          additionalProperties:
            $ref: '#/components/schemas/MessageArray'
  responses:
    ServiceResponse:
      content:
        application/json:
          schema:
            oneOf:
            - $ref: '#/components/schemas/Service'
            - $ref: '#/components/schemas/ServiceView'
      description: ""
  securitySchemes:
    jwt-bearer:
      flows:
        clientCredentials:
          tokenUrl: https://keycloak.example.com/auth/realms/microcks/protocol/openid-connect/token
          refreshUrl: https://keycloak.example.com/auth/realms/microcks/protocol/openid-connect/token
          scopes:
            user: ""
            manager: ""
            admin: ""
      type: oauth2
      description: JWT Bearer acquired using OAuth 2 Authentication flow or Direct
        Access Grant
security:
- jwt-bearer: []
...